requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Authentication
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0

# Pydantic
pydantic>=2.5.0
pydantic-settings>=2.1.0

# HTTP client
httpx>=0.26.0

# Encryption
cryptography>=42.0.0

# Templates
jinja2>=3.1.2

# MCP
mcp>=1.0.0

# SSE
sse-starlette>=1.6.0

# Telegram (MTProto)
telethon>=1.37.0
